# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Bayt Al Fann
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Unauthorized
  /auth/me:
    get:
      operationId: getAdminMe
      tags: [auth]
      summary: Get current logged in admin
      responses:
        "200":
          description: Logged in admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
  /artworks:
    get:
      operationId: getArtworks
      tags: [artworks]
      summary: List all artworks
      parameters:
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: isFeatured
          in: query
          required: false
          schema:
            type: boolean
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of artworks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artwork"
    post:
      operationId: createArtwork
      tags: [artworks]
      summary: Create a new artwork (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtworkInput"
      responses:
        "201":
          description: Artwork created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
  /artworks/{id}:
    get:
      operationId: getArtworkById
      tags: [artworks]
      summary: Get artwork detail by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artwork details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "404":
          description: Artwork not found
    put:
      operationId: updateArtwork
      tags: [artworks]
      summary: Update artwork details (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtworkInput"
      responses:
        "200":
          description: Artwork updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
    delete:
      operationId: deleteArtwork
      tags: [artworks]
      summary: Delete an artwork (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artwork deleted
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category deleted
  /blog:
    get:
      operationId: getBlogPosts
      tags: [blog]
      summary: List all blog posts
      responses:
        "200":
          description: List of blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Blog post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{id}:
    get:
      operationId: getBlogPostById
      tags: [blog]
      summary: Get blog post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    put:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "200":
          description: Blog post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blog post deleted
  /exhibitions:
    get:
      operationId: getExhibitions
      tags: [exhibitions]
      summary: List all exhibitions
      responses:
        "200":
          description: List of exhibitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exhibition"
    post:
      operationId: createExhibition
      tags: [exhibitions]
      summary: Create an exhibition (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExhibitionInput"
      responses:
        "201":
          description: Exhibition created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exhibition"
  /exhibitions/{id}:
    get:
      operationId: getExhibitionById
      tags: [exhibitions]
      summary: Get exhibition by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exhibition details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exhibition"
    put:
      operationId: updateExhibition
      tags: [exhibitions]
      summary: Update an exhibition (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExhibitionInput"
      responses:
        "200":
          description: Exhibition updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exhibition"
    delete:
      operationId: deleteExhibition
      tags: [exhibitions]
      summary: Delete an exhibition (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exhibition deleted
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List all orders (Admin)
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a checkout order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrderById
      tags: [orders]
      summary: Get order details by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    put:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdateInput"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /commissions:
    get:
      operationId: getCommissionRequests
      tags: [commissions]
      summary: List all commissions (Admin)
      responses:
        "200":
          description: List of commission requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommissionRequest"
    post:
      operationId: createCommissionRequest
      tags: [commissions]
      summary: Submit a commission request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommissionInput"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionRequest"
  /commissions/{id}:
    put:
      operationId: updateCommissionRequestStatus
      tags: [commissions]
      summary: Accept or Reject commission request (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, accepted, rejected]
              required: [status]
      responses:
        "200":
          description: Request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionRequest"
  /contact:
    get:
      operationId: getContactMessages
      tags: [contact]
      summary: List all contact messages (Admin)
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
    post:
      operationId: submitContactMessage
      tags: [contact]
      summary: Send contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Message submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
  /contact/{id}/archive:
    put:
      operationId: archiveContactMessage
      tags: [contact]
      summary: Archive a contact message (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        "201":
          description: Subscribed successfully
  /testimonials:
    get:
      operationId: getApprovedTestimonials
      tags: [testimonials]
      summary: List approved testimonials
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: submitTestimonial
      tags: [testimonials]
      summary: Submit a testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Testimonial submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
...
  /testimonials/{id}/approve:
    put:
      operationId: approveTestimonial
      tags: [testimonials]
      summary: Approve or reject testimonial (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isApproved:
                  type: boolean
              required: [isApproved]
      responses:
        "200":
          description: Testimonial status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get website settings
      responses:
        "200":
          description: Public site settings object
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update website settings (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
      responses:
        "200":
          description: Settings updated
...
components:
  schemas:
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        admin:
          $ref: "#/components/schemas/AdminUser"
      required: [token, admin]
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
      required: [id, username, email]
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image:
          type: string
      required: [id, name, slug]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        image:
          type: string
      required: [name, slug]
...
    Artwork:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: integer
        dimensions:
          type: string
        medium:
          type: string
        availability:
          type: string
        rating:
          type: integer
        categoryId:
          type: integer
        certificateNumber:
          type: string
        isFeatured:
          type: boolean
        isLimitedEdition:
          type: boolean
        totalEditions:
          type: integer
        availableEditions:
          type: integer
        images:
          type: array
          items:
            $ref: "#/components/schemas/ArtworkImage"
        createdAt:
          type: string
      required: [id, title, slug, description, price, dimensions, medium, availability, isFeatured, isLimitedEdition]
    ArtworkInput:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: integer
        dimensions:
          type: string
        medium:
          type: string
        availability:
          type: string
        categoryId:
          type: integer
        certificateNumber:
          type: string
        isFeatured:
          type: boolean
        isLimitedEdition:
          type: boolean
        totalEditions:
          type: integer
        availableEditions:
          type: integer
        images:
          type: array
          items:
            type: string
      required: [title, slug, description, price, dimensions, medium]
...
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        excerpt:
          type: string
        readTime:
          type: string
        coverImage:
          type: string
        author:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required: [id, title, slug, content, excerpt, readTime, coverImage, author, status, createdAt]
...